scripts/tonnel/requirements.txt
curl_cffi
psycopg2-binary
python-dotenv